requirements.txt
python-telegram-bot==20.*
fasttext==0.9.2